Docker image/requirements.txt
selenium==4.25.0
webdriver-manager==4.0.2
psycopg2-binary==2.9.9
selenium-stealth==1.0.6
